git-svn-id: http://svn.fortytwo.net/projects@188 4bf1a796-0a03-0410-9b82-c499c1e613c7

+++ troika/phase2/trunk/src/api/util/Array.c
@@ -57,12 +57,12 @@ static void **
 buffer_copy( Array *a )
 {
     void **buffer;
-    int size = a->buffer_size * sizeof( void* );
+    size_t size = a->buffer_size * sizeof( void* );
 
     if ( ( buffer = malloc( size ) ) )
         memcpy( buffer, a->buffer, size );
 
-    return buffer;
+    return buffer /*@null@*/;
 }
